Threading!
I was able to get threading working after about ...a million revisions. Its included in the code now, and is built in a way that either the API will run, or the Kafka consumer will run. It doesnt appear -yet- that I can have them both running at the same time. More research is needed